requirements.txt
mlflow>=2.21.2
streamlit==1.44.1
databricks-sdk>=0.60.0
Pillow>=10.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pandas>=1.5.0
psycopg>=3.2.0
PyPDF2>=3.0.0
